Renaming and docstrings.
